[serve][llm] Move in-process tokenizer under routing_policies/kv_aware
Pure relocation of the tokenizer module (and its unit test) from
core/ingress to routing_policies/kv_aware, updating import paths. No
logic change; the vLLM-specific rewrite follows in the next commit.
