+++ python/ray/llm/tests/serve/cpu/deployments/routers/kv/test_kv_aware_router.py
@@ -23,11 +23,13 @@ from ray.llm._internal.serve.core.ingress.builder import (
     LLMServingArgs,
     build_openai_app,
 )
-from ray.llm._internal.serve.core.ingress.tokenizer import REQUEST_TOKEN_IDS_KWARG
 from ray.llm._internal.serve.routing_policies.kv_aware.kv_aware_actor import (
     KV_ROUTER_ACTOR_NAME,
     KVRouterActor,
     get_worker_id,
+)
+from ray.llm._internal.serve.routing_policies.kv_aware.tokenizer import (
+    REQUEST_TOKEN_IDS_KWARG,
 )
 from ray.serve._private.common import (
     REPLICA_ID_FULL_ID_STR_PREFIX,
