+++ python/ray/llm/tests/serve/cpu/deployments/routers/kv/test_tokenizer.py
@@ -18,7 +18,10 @@ from ray.llm._internal.serve.core.ingress.builder import (
     build_openai_app,
 )
 from ray.llm._internal.serve.core.ingress.router import LLMRouter
-from ray.llm._internal.serve.core.ingress.tokenizer import TokenizeError, Tokenizer
+from ray.llm._internal.serve.routing_policies.kv_aware.tokenizer import (
+    TokenizeError,
+    Tokenizer,
+)
 from ray.serve.experimental.round_robin_router import RoundRobinRouter
 from ray.serve.llm.request_router import KVAwareRouter
